mm : add lowmem detection logic
From: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>

Final purpose of this patch is for improving oom/memoy shortage
detection
better. In general there are OOM cases that lowmem is exhausted. What
this lowmem means is determined by the situation, but in general,
limited amount of memory for some special use is lowmem.

This patch adds an integer lowmem_zone, which is initialized to -1.
If zone_idx(zone) <= lowmem_zone, the zone is lowmem.

This patch uses simple definition that the zone for special use is the
lowmem.
Not taking the amount of memory into account.

For example,
- if HIGHMEM is used, NORMAL is lowmem.
- If the system has both of NORMAL and DMA32, DMA32 is lowmem.
- When the system consists of only one zone, there are no lowmem.

This will be used for lowmem accounting per mm_struct and its
information
will be used for oom-killer.

Q: Why you don't use policy_zone ?
A: It's for NUMA only. I want to use unified approach for detecting
lowmem.
And policy_zone sounds like "for mempolicy"..

Concerns or TODO:
- Now, we have polizy_zone if CONFIG_NUMA=y. Maybe we can make it as
or remove it. But this itself should be done in other patch.

Changelog: 2009/12/14
- no change.
Changelog: 2009/12/09
- stop using policy_zone and use unified definition on each config.

Reviewed-by: Minchan Kim <minchan.kim@gmail.com>
Signed-off-by: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>